# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AgentoGo API",
        "description": "API REST multi-tenant con autenticación, auditoría y gestión de negocios",
        "contact": {
            "name": "AgentoGo Support",
            "email": "support@agentogo.com"
        },
        "version": "0.3.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor de desarrollo"
        }
    ],
    "paths": {
...
        "/businesses/{business_id}/employees": {
            "get": {
                "tags": [
                    "Employee"
                ],
                "summary": "Listar empleados del negocio",
                "operationId": "bd6fa756b336e9467fab75d7ea7182e3",
                "responses": {
                    "200": {
                        "description": "Lista de empleados con paginación"
                    },
                    "403": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error403"
                                },
                                "example": {
                                    "code": 403,
                                    "message": "No tiene permisos para realizar esta acción."
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Employee"
                ],
                "summary": "Crear un empleado",
                "operationId": "55d00ff4444accb947d9329316bccda5",
                "responses": {
                    "201": {
                        "description": "Empleado creado"
                    },
                    "403": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error403"
                                },
                                "example": {
                                    "code": 403,
                                    "message": "No tiene permisos para realizar esta acción."
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validación fallida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                },
                                "example": {
                                    "message": "The given data was invalid.",
                                    "errors": {
                                        "name": [
                                            "El nombre es obligatorio."
                                        ],
                                        "email": [
                                            "El email debe ser una dirección válida."
                                        ]
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/businesses/{business_id}/employees/{employee_id}": {
            "get": {
                "tags": [
                    "Employee"
                ],
                "summary": "Obtener un empleado",
                "operationId": "95dc1d3e7a36b8b99317cda2b0021c07",
                "responses": {
                    "200": {
                        "description": "Detalle del empleado"
                    },
                    "403": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error403"
                                },
                                "example": {
                                    "code": 403,
                                    "message": "No tiene permisos para realizar esta acción."
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Employee"
                ],
                "summary": "Actualizar un empleado",
                "operationId": "a39acddc014ee232fa268a1f9e269bf0",
                "responses": {
                    "200": {
                        "description": "Empleado actualizado"
                    },
                    "403": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error403"
                                },
                                "example": {
                                    "code": 403,
                                    "message": "No tiene permisos para realizar esta acción."
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validación fallida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                },
                                "example": {
                                    "message": "The given data was invalid.",
                                    "errors": {
                                        "name": [
                                            "El nombre es obligatorio."
                                        ],
                                        "email": [
                                            "El email debe ser una dirección válida."
                                        ]
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Employee"
                ],
                "summary": "Eliminar un empleado",
                "operationId": "63bceeb1b13d316d4dea53dcd95f9ca6",
                "responses": {
                    "204": {
                        "description": "Empleado eliminado"
                    },
                    "403": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error403"
                                },
                                "example": {
                                    "code": 403,
                                    "message": "No tiene permisos para realizar esta acción."
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Error403": {
                "properties": {
                    "code": {
                        "type": "integer",
                        "example": 403
                    },
                    "message": {
                        "type": "string",
                        "example": "No autorizado"
                    }
                },
                "type": "object",
                "example": {
                    "code": 403,
                    "message": "No tiene permisos para realizar esta acción."
                }
            },
            "ValidationError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "description": "Mapa de errores por campo",
                        "type": "object"
                    }
                },
                "type": "object",
                "example": {
                    "message": "The given data was invalid.",
                    "errors": {
                        "name": [
                            "El campo nombre es requerido."
                        ],
                        "email": [
                            "El email debe ser una dirección válida."
                        ]
                    }
                }
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token de autenticación Sanctum",
                "bearerFormat": "token",
                "scheme": "bearer"
            }
        }
    },
...
}